refactor: split ConfigController into handler and service classes
